backup_20260225_211753/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
python-dotenv==1.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
python-multipart==0.0.9
redis==5.0.8
python-telegram-bot==21.6
psycopg2-binary==2.9.9
flask>=3.0.0
aiohttp
qrcode[pil]
httpx
pydantic==2.10.6
pydantic-settings==2.7.1

prometheus-fastapi-instrumentator
json-logging
json-logging
